requirements.txt
alembic==1.13.1
and==66.0.2
aniso8601==9.0.1
annotated-types==0.6.0
anyio==3.7.1
apturl==0.5.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.15.0
bcrypt==3.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
Brlapi==0.8.3
certifi==2020.6.20
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.4
comm==0.2.2
command-not-found==0.3
contourpy==1.2.0
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
DateTime==5.4
dbus-python==1.2.18
debugpy==1.8.1
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.7
distro==1.7.0
distro-info==1.1+ubuntu0.2
duplicity==0.8.21
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.105.0
fasteners==0.14.1
fastjsonschema==2.19.1
filelock==3.13.1
Flask==3.0.1
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-Mail==0.10.0
Flask-Migrate==4.0.5
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
fonttools==4.47.0
fqdn==1.5.1
future==0.18.2
greenlet==3.0.3
gyp==0.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.20.2
httpx==0.27.0
idna==3.3
importlib-metadata==4.6.4
ipykernel==6.29.4
ipython==8.24.0
ipywidgets==8.1.2
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.3
json5==0.9.25
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.8
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.1
jupyterlab_widgets==3.0.10
keyring==23.5.0
kiwisolver==1.4.5
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.4
matplotlib==3.8.2
matplotlib-inline==0.1.7
mistune==3.0.2
monotonic==1.6
more-itertools==8.10.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
netifaces==0.11.0
notebook==7.1.3
notebook_shim==0.2.4
numpy==1.26.3
oauthlib==3.2.0
olefile==0.46
overrides==7.7.0
packaging==23.2
pandocfilters==1.5.1
paramiko==2.9.3
parso==0.8.4
pexpect==4.8.0
Pillow==9.0.1
pipenv==2023.11.15
platformdirs==3.11.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==3.12.4
psutil==5.9.8
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pycairo==1.20.1
pycparser==2.22
pycups==2.0.1
pydantic==2.5.2
pydantic_core==2.14.5
Pygments==2.18.0
PyGObject==3.42.1
pyinotify==0.9.6
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==26.0.3
qtconsole==5.5.2
QtPy==2.4.1
referencing==0.35.1
reportlab==3.6.8
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.0
SecretStorage==3.3.1
Send2Trash==1.8.3
simplejson==3.17.6
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.25
SQLAlchemy-serializer==1.4.1
stack-data==0.6.3
starlette==0.27.0
systemd-python==234
terminado==0.18.1
tinycss2==1.3.0
tomli==2.0.1
tornado==6.4
traitlets==5.14.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.9.0
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
uri-template==1.3.0
urllib3==1.26.5
usb-creator==0.3.7
uvicorn==0.24.0.post1
virtualenv==20.24.6
virtualenv-clone==0.3.0
wadllib==1.3.6
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.1
widgetsnbextension==4.0.10
xdg==5
xkit==0.0.0
zipp==1.0.0
zope.interface==6.1
